resume_ai/requirements.txt
flask
scikit-learn
pandas
nltk
python-dotenv
sentence-transformers
pymongo
PyPDF2
streamlit
fastapi
uvicorn
joblib
pydantic
